'mybatis'

+++ src/main/java/com/olliego/community/mapper/QuestionMapper.java
@@ -18,10 +18,10 @@ public interface QuestionMapper {
     Integer count();
 
     @Select("select * from question where creator = #{userId} limit #{offset},#{size}")
-    List<Question> listByUserId(@Param(value = "userId") Integer userId, @Param(value = "offset") Integer offset, @Param(value = "size") Integer size);
+    List<Question> listByUserId(@Param(value = "userId") Long userId, @Param(value = "offset") Integer offset, @Param(value = "size") Integer size);
 
     @Select("select count(1) from question where creator = #{userId}")
-    Integer countByUserId(@Param(value = "userId") Integer userId);
+    Integer countByUserId(@Param(value = "userId") Long userId);
 
     @Select("select * from question where id = #{id}")
     Question getById(@Param("id") Integer id);
